Add per-vehicle reminder muting and float odometer support.

+++ backend/app/models/service_log.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import date as dt_date
 
-from sqlalchemy import Column, Date, Float, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import Column, Date, Float, ForeignKey, Index, String, Text
 from sqlalchemy.dialects.postgresql import UUID, ARRAY
 from sqlalchemy.orm import relationship
 
@@ -16,12 +16,12 @@ class ServiceLog(Base, TimestampMixin):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     vehicle_id = Column(UUID(as_uuid=True), ForeignKey("vehicles.id"), nullable=False)
     date = Column(Date, nullable=False)
-    odometer = Column(Integer, nullable=False)
+    odometer = Column(Float, nullable=False)
     service_center = Column(String, nullable=True)
     total_cost = Column(Float, nullable=False, default=0)
     services_done = Column(ARRAY(String), nullable=False)
     next_service_date = Column(Date, nullable=True)
-    next_service_odometer = Column(Integer, nullable=True)
+    next_service_odometer = Column(Float, nullable=True)
     notes = Column(Text, nullable=True)
 
     vehicle = relationship("Vehicle", back_populates="service_logs")
